requirements.txt
Pillow>=10.0.0
piexif>=1.1.3
tqdm>=4.65.0
hachoir>=3.2.0 